requirements.txt
Django==2.2.2
django-crispy-forms==1.7.2
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
